@robodoo 2025-03-24
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/203182.png)](https://mergebot.odoo.com/odoo/odoo/pull/203182)